Restructure plugin editor folders
Nest folders under `Editor` in preparation for adding asmdefs. This must be done because there is a biredctional dependency between `Core/Editor` and `Tools/*/Editor` folders which can only be worked around using the "assembly reference" asset which is unsupported in Unity 2018.

Note that having an `Editor` folder at the root is also the official suggested package layout (https://docs.unity3d.com/Manual/cus-layout.html).

I've excluded `Example/Editor` from this because the examples are self-contained and arguably should not be included in the distributed package.